requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
gunicorn==21.2.0
idna==2.10
packaging==24.2
Pillow==9.0.1
pycparser==2.22
PyJWT==2.10.1
pytz==2025.2
requests==2.25.1
sqlparse==0.5.3
urllib3==1.26.20